[💫 Feature] - Battle Class 분리, 디버그 종료 시 Reward 호출
- Battle class => LeftWindow, RightWindow, BugSolution 총 4개의 스크립트로 분리
- 배틀 종료 시 Reward Class 호출
- 호출 시 Data.BattleSuccess bool 값 확인
