+++ TeamDungeon/HexaCoreVillage/Dungeon/LoggingText.cs
@@ -1,0 +1,11 @@
+namespace HexaCoreVillage.Dungeon
+{
+    public class LoggingText
+    {
+        public List<string>? ModifyVariables { get; set; }
+        public List<string>? FixTypos { get; set; }
+        public List<string>? ModifyScope { get; set; }
+        public List<string>? EditPath { get; set; }
+        public List<string>? CheckMemory { get; set; }
+    }
+}
